Minor edits to the speculative Jinja template format...
All these edits presume that the actual variable names are as detailed.
They also presume data not present in the current response: the origin
point location, the individual trip status, departure times, and
arrival times. It also presupposes that we can do the inner steps loop
to get all the step logos...