Added a transit module to wrap the Bing API
I ported the code from the web.py module for
processing the raw Bing API output and wired it
to an express handler.

The 0.0.1 version of the bing module was 
incorrectly sending local time rather than UTC to
the Bing Maps API resulting in incorrect travel 
times. This was fixed in the 0.0.2 version.